changed int by size_t in render

+++ src/render.c
@@ -6,7 +6,7 @@
 /*   By: omimouni <omimouni@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/20 02:39:09 by omimouni          #+#    #+#             */
-/*   Updated: 2020/12/20 02:44:48 by omimouni         ###   ########.fr       */
+/*   Updated: 2020/12/20 03:22:28 by omimouni         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 void	render(t_conf *conf)
 {
-	int	i;
-	int	j;
+	size_t	i;
+	size_t	j;
 
 	i = 0;
 	while (i < 700)
